plugins/demo/requirements.txt
cloudify-common==4.6
